+++ react/types/index.ts
@@ -4,6 +4,12 @@ export interface BasicPriceProps {
   message: string
   markers: string[]
 }
+
+export interface PriceRangeProps {
+  message: string
+  noRangeMessage: string
+  markers: string[]
+}
 export interface StorefrontFC<P = {}> extends FC<P> {
   getSchema?(props: P): object
   schema?: object