+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `ListPriceRange` and `SellingPriceRange` component.
+
+### Changed
+- Made `ListPrice` render null if it is the same as `SellingPrice`.
 
 ## [0.4.1] - 2020-03-18
 ### Changed